decomp + native port: implement 30+ functions, web port gameplay fixes, RTT docs
Decomp:
- Implement functions across Ribbon, Part, VelocityBuffer, SoftParticleBuffer,
  DOFProc_NG, Lit_NG, ShaderProgram, AmbientOcclusion, Font, MeshDeform,
  BinkIntegration, MemTrack, NetCacheLoader, PhysicsVolume, Spotlight,
  SpotlightDrawer, Instance, FriendsListJobs, FreestyleMoveRecorder,
  MoveAsyncDetector, MoveDir, ByteGrinder, EncryptXTEA, File, Text,
  UIListSlot, DataPointMgr, HamStorePanel
- Fix match percentages in kdTree, Endian, PostProc_NG, Cam, Rnd_NG

Web port gameplay:
- Fix "movemgr not function or object" error spam (lightweight stub in main_web)
- Fix "get_disable_freestyle" unhandled (WebProfileMgr handler)
- Fix "video_recorder.srec" missing (stub in HamDirector venue load)
- Fix MoveMgr assert failures (null guards with HX_NATIVE early returns)
- Fix web build hanging on song load (use stub instead of heavy MoveMgr::Init)
- Fix character animations not advancing (force Showing during HamCharacter::Poll)
- Fix Draw3DChars skipping rendering on native (add HX_NATIVE draw path)

Docs:
- New: docs/native/RENDER_TO_TEXTURE.md — RTT architecture, consumer status
- Update RENDERING_SYSTEM.md, GRAPHICS_SYSTEM_DESIGN.md with RTT status
- Update native port status, rendering gaps, decomp gaps

+++ src/system/synth/ByteGrinder.h
@@ -5,11 +5,11 @@ class ByteGrinder {
 public:
     ByteGrinder() {}
     virtual ~ByteGrinder() {} // generic dtor
-    void GrindArray(int, int, unsigned char *, int, int);
+    void GrindArray(long, long, unsigned char *, int, int);
     void Init();
     void HvDecrypt(unsigned char *, unsigned char *, int);
 
 private:
-    unsigned int pickOneOf32A(bool, int);
-    unsigned int pickOneOf32B(bool, int);
+    unsigned long pickOneOf32A(bool, long);
+    unsigned long pickOneOf32B(bool, long);
 };
